refactor(chat): isolate lark connection api
