+++ loopx/chat_lark_api.py
@@ -1,0 +1,250 @@
+from __future__ import annotations
+
+import subprocess
+from pathlib import Path
+from typing import Any
+from urllib.parse import parse_qs, urlparse
+
+from .control_plane.runtime.runtime_projection_route import resolve_goal_source_runtime_route
+from .extensions.lark.goal_channel import (
+    default_goal_channel_binding_path,
+    default_goal_channel_target_path,
+)
+from .extensions.lark.goal_topic_connections import (
+    connect_lark_goal_topic,
+    disconnect_lark_goal_topic,
+    list_lark_apps,
+    list_lark_connections,
+    list_lark_group_chats,
+)
+from .extensions.lark.presentation.kanban import CommandRunner, default_subprocess_runner
+from .history import load_registry
+from .paths import resolve_runtime_root
+from .registry import registry_goals
+from .repository_identity import normalize_repository_identity
+
+
+def _compact_text(value: Any, *, limit: int = 600) -> str:
+    return " ".join(str(value or "").split())[:limit].strip()
+
+
+def _default_git_runner(args: list[str]) -> dict[str, Any]:
+    try:
+        completed = subprocess.run(
+            args,
+            capture_output=True,
+            check=False,
+            text=True,
+            timeout=10,
+        )
+    except (OSError, subprocess.TimeoutExpired):
+        return {"returncode": 1, "stdout": ""}
+    return {"returncode": completed.returncode, "stdout": completed.stdout}
+
+
+def build_goal_repository_contexts(
+    *,
+    registry: dict[str, Any],
+    git_runner: Any = _default_git_runner,
+) -> list[dict[str, Any]]:
+    """Return local-API Goal repository labels without checkout paths."""
+
+    rows: list[dict[str, Any]] = []
+    for goal in registry_goals(registry):
+        goal_id = str(goal.get("id") or "").strip()
+        repo_text = str(goal.get("repo") or "").strip()
+        if not goal_id or not repo_text:
+            continue
+        project = Path(repo_text).expanduser().resolve()
+        remote_result = git_runner(
+            ["git", "-C", str(project), "config", "--get", "remote.origin.url"]
+        )
+        remote = str(remote_result.get("stdout") or "").strip()
+        try:
+            identity = normalize_repository_identity(remote)
+        except ValueError:
+            identity = f"loopx:{goal_id}"
+        branch_result = git_runner(
+            ["git", "-C", str(project), "branch", "--show-current"]
+        )
+        branch = _compact_text(branch_result.get("stdout"), limit=160)
+        label = identity.split("/", 1)[1] if identity.startswith("git:") and "/" in identity else goal_id
+        rows.append(
+            {
+                "goal_id": goal_id,
+                "repository": {
+                    "branch": branch,
+                    "identity": identity,
+                    "label": label,
+                    "read_only": True,
+                },
+            }
+        )
+    return rows
+
+
+class LarkChatRequestMixin:
+    """Loopback HTTP handlers for Goal repository and Lark topic connections."""
+
+    server: Any
+    path: str
+
+    def _goal_channel_context(self, goal_id: str) -> tuple[dict[str, Any], Path]:
+        registry, _goal = self._registry_and_goal(goal_id)
+        route = resolve_goal_source_runtime_route(
+            registry_path=self.server.registry_path,
+            goal_id=goal_id,
+            registry=registry,
+        )
+        source_registry_path = Path(str(route["source_registry"]))
+        source_registry = (
+            registry
+            if source_registry_path.expanduser().resolve()
+            == self.server.registry_path.expanduser().resolve()
+            else load_registry(source_registry_path)
+        )
+        return source_registry, default_goal_channel_binding_path(source_registry_path)
+
+    def _goal_channel_target_path(self) -> Path:
+        registry = load_registry(self.server.registry_path)
+        runtime_root = resolve_runtime_root(
+            registry,
+            self.server.runtime_root_override,
+            registry_path=self.server.registry_path,
+        )
+        return default_goal_channel_target_path(runtime_root)
+
+    def _lark_runner(self) -> CommandRunner:
+        return getattr(self.server, "lark_runner", default_subprocess_runner)
+
+    def _lark_binding_paths(self, registry: dict[str, Any]) -> dict[str, Path]:
+        paths: dict[str, Path] = {}
+        for goal in registry_goals(registry):
+            goal_id = str(goal.get("id") or "").strip()
+            if not goal_id:
+                continue
+            try:
+                route = resolve_goal_source_runtime_route(
+                    registry_path=self.server.registry_path,
+                    goal_id=goal_id,
+                    registry=registry,
+                )
+                source_registry_path = Path(str(route["source_registry"]))
+                if source_registry_path.parent.name == ".loopx":
+                    paths[goal_id] = default_goal_channel_binding_path(source_registry_path)
+            except (OSError, ValueError):
+                continue
+        return paths
+
+    def _goal_contexts(self) -> None:
+        registry = load_registry(self.server.registry_path)
+        self._send_json(
+            {
+                "ok": True,
+                "schema_version": "loopx_chat_goal_contexts_v0",
+                "goals": build_goal_repository_contexts(registry=registry),
+            }
+        )
+
+    def _lark_apps(self) -> None:
+        self._send_json(
+            {
+                "ok": True,
+                "schema_version": "loopx_lark_apps_v0",
+                "apps": list_lark_apps(runner=self._lark_runner()),
+            }
+        )
+
+    def _lark_chats(self) -> None:
+        query = parse_qs(urlparse(self.path).query)
+        app_ref = _compact_text(query.get("app_ref", [""])[0], limit=100)
+        keyword = _compact_text(query.get("query", [""])[0], limit=120)
+        if not app_ref:
+            self._send_error("app_ref is required", status=400, error_code="lark_app_required")
+            return
+        try:
+            chats = list_lark_group_chats(
+                app_ref=app_ref,
+                query=keyword or None,
+                runner=self._lark_runner(),
+            )
+        except ValueError as exc:
+            self._send_error(str(exc), status=400, error_code="invalid_lark_app")
+            return
+        self._send_json(
+            {
+                "ok": True,
+                "schema_version": "loopx_lark_group_chats_v0",
+                "chats": chats,
+            }
+        )
+
+    def _lark_connections(self) -> None:
+        registry = load_registry(self.server.registry_path)
+        self._send_json(
+            {
+                "ok": True,
+                "schema_version": "loopx_lark_goal_topic_connections_v0",
+                "connections": list_lark_connections(
+                    registry=registry,
+                    target_path=self._goal_channel_target_path(),
+                    binding_paths=self._lark_binding_paths(registry),
+                ),
+            }
+        )
+
+    def _lark_connect(self) -> None:
+        try:
+            body = self._read_json()
+            allowed = {"app_ref", "chat_id", "chat_name", "execute", "goal_id", "incoming_mode"}
+            if set(body) - allowed:
+                raise ValueError("unknown Lark connection field")
+            goal_id = _compact_text(body.get("goal_id"), limit=160)
+            app_ref = _compact_text(body.get("app_ref"), limit=100)
+            chat_id = _compact_text(body.get("chat_id"), limit=160)
+            chat_name = _compact_text(body.get("chat_name"), limit=120)
+            incoming_mode = _compact_text(body.get("incoming_mode"), limit=40) or "mentions"
+            if not goal_id or not app_ref or not chat_id or not chat_name:
+                raise ValueError("goal_id, app_ref, chat_id, and chat_name are required")
+            registry, binding_path = self._goal_channel_context(goal_id)
+            packet = connect_lark_goal_topic(
+                registry=registry,
+                goal_id=goal_id,
+                target_path=self._goal_channel_target_path(),
+                binding_path=binding_path,
+                app_ref=app_ref,
+                chat_id=chat_id,
+                chat_name=chat_name,
+                incoming_mode=incoming_mode,
+                execute=body.get("execute") is True,
+                runner=self._lark_runner(),
+            )
+        except ValueError as exc:
+            self._send_error(str(exc), status=400, error_code="invalid_lark_connection")
+            return
+        except Exception:
+            self._send_error(
+                "the Lark connection operation failed before a verified provider receipt",
+                status=400,
+                error_code="provider_api_failed",
+            )
+            return
+        if not packet.get("ok"):
+            packet["error"] = _compact_text(
+                packet.get("public_summary") or packet.get("blocker") or "Lark connection failed"
+            )
+        self._send_json(packet, status=200 if packet.get("ok") else 400)
+
+    def _lark_disconnect(self) -> None:
+        query = parse_qs(urlparse(self.path).query)
+        goal_id = _compact_text(query.get("goal_id", [""])[0], limit=160)
+        if not goal_id:
+            self._send_error("goal_id is required", status=400, error_code="goal_required")
+            return
+        try:
+            _registry, binding_path = self._goal_channel_context(goal_id)
+            packet = disconnect_lark_goal_topic(binding_path=binding_path, goal_id=goal_id)
+        except (OSError, ValueError) as exc:
+            self._send_error(str(exc), status=400, error_code="invalid_lark_connection")
+            return
+        self._send_json(packet)
